pyproject.toml
[tool.poetry]
name = "labelstudio2rttm"
version = "0.1.0"
description = "LabelStudio to RTTM annotation converter"
authors = ["Luca Bondi <luca.bondi@us.bosch.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
defopt = "^6.4.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
